feat: Add DonationsTile
